This PR contains changes to use the native Avro Date representation to support a more direct mapping to parquet and avro files without having to remap all dates to their ISO string representation.
Avro dates:
http://avro.apache.org/docs/current/spec.html#Date

The date logical type represents a date within the calendar, with no reference to a particular time zone or time of day.

A date logical type annotates an Avro int, where the int stores the number of days from the unix epoch, 1 January 1970 (ISO calendar).

The following schema represents a date:

{
  "type": "int",
  "logicalType": "date"
}